# backend/v2/openapi.yaml
openapi: 3.0.3
info:
  title: AI Skill Management Platform
  version: 2.0.0
  description: Multi-tenant SaaS for creating, managing, assigning, and executing AI skills
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/users:
    get:
      summary: List organization users (paginated)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated users list
  /api/users/{user_id}:
    get:
      summary: Get user profile
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User detail
    patch:
      summary: Update user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                role: { type: string, enum: [OWNER, ADMIN, MEMBER, VIEWER] }
                status: { type: string, enum: [active, inactive] }
      responses:
        '200':
          description: User updated
    delete:
      summary: Deactivate user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User deactivated
  /api/users/{user_id}/skills:
    get:
      summary: Get skills assigned to user
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User skills list
    post:
      summary: Assign skills to user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_ids]
              properties:
                skill_ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '201':
          description: Skills assigned
  /api/users/{user_id}/skills/{skill_id}:
    delete:
      summary: Remove skill from user (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: skill_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Skill removed
  /api/users/invite:
    post:
      summary: Invite user by email (admin)
      tags: [Users]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, default: MEMBER }
      responses:
        '201':
          description: Invitation created
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: A skill with this name already exists
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
...